refactor(export): split the row types out of model.rs for the size ratchet
CI caught what I did not: model.rs grew to 555 lines against its recorded 468
budget, and the ratchet's instruction is to shrink or split rather than raise
the number.

model_types.rs now holds the crate's public data shape (PropValue,
PropertySet, QuantityValue, QuantitySet, EntityRow, ExportModel), which is the
natural seam: every exporter and the Python binding read those types, and none
of them cares how the streaming pass fills them in. Re-exported from model.rs,
so the public API is byte-identical. 445 lines now, and the new file is under
the 400 limit so it needs no allowlist row of its own.

My mistake was verifying at the wrong scope. I ran -p ifc-lite-export and
workspace clippy, neither of which runs the ratchet, which lives in
ifc-lite-processing's test target. Re-ran cargo test --workspace this time,
with the staged fixtures symlinked in so the geometry suites actually execute:
zero failures across the workspace.
